fix(grading): widen peakMem to Long to prevent silent narrowing overflow
The initial extract in f7e813a narrowed AppViewModel.peakMem (Long) to Int at the GradingInput boundary. Safe in current heap ranges (<2 GB MB) but a latent overflow if the app ever targets larger heaps. This commit keeps Long end-to-end; no behavior change in the penalty thresholds (2000 MB / 1500 MB) which auto-promote.

+++ src/main/kotlin/com/gameperf/desktop/core/grading/FinalScoreCalculator.kt
@@ -27,7 +27,7 @@ data class GradingInput(
     val totalJank: Long,
     val finalElapsed: Double,
     val totalStutter: Int,
-    val peakMem: Int,
+    val peakMem: Long,
     val maxTempCpu: Double,
     val avgCpu: Int,
 )
